Get PSI metric from hypervisor
Change the resource monitor for CPU to fetch the presure stall
information (PSI) from the hypervisor. Along with the other
CPU stats it can be saved as a metric for later use by components
like the scheduler. Such a metric could help relief VM steal time.

As there's no libvirt support for fetching PSI, the way to do it
for now is to perform a read on the kernel interface found at
`/proc/pressure/cpu`.

Signed-off-by: Raul Moldovan <rmoldovan3@bloomberg.net>